refactor(messaging): relocate repo interfaces to application/usecases (DIP)
v0.162.1 polish Item 1 — closes one of 4 Tier 2 carry-forwards from v0.162.0.

ConversationRepository + MessageRepository moved from
`internal/modules/messaging/domain/repositories/` to
`internal/modules/messaging/application/usecases/repository_interfaces.go`
per Clean Architecture DIP gate (CLAUDE.md: "Repository interfaces — в
пакете-потребителе (`usecase/`), НЕ в `domain/`"). Mirrors v0.157.1
curriculum + v0.163.1 announcements precedent.

Zero behavior change:
- Interfaces moved verbatim, constructor signatures preserved
  via Go structural typing (concrete *RepositoryPG satisfies the
  relocated interface).
- 2 PG impls (conversation_repository_pg.go, message_repository_pg.go)
  switch their named import to messagingUsecases.
- main.go drops messagingRepositories alias, switches
  newMessagingAccessChecker arg type to messagingUsecases.ConversationRepository.
- Filter DTOs (ConversationFilter, MessageFilter) stay in domain/entities/
  — they were never in domain/repositories/.
- No domain-level sentinels were in domain/repositories/ (messaging
  entity-level invariants live in domain/entities/, e.g. ErrNotParticipant),
  so the package becomes empty and is removed.

All messaging tests pass post-relocation.

Refs #311
